+++ src/Controller/BaladesController.php
@@ -9,6 +9,7 @@ use App\Entity\Balades;
 use App\Form\BaladesType;
 use App\Repository\BaladesRepository;
 use App\Repository\ActivitesRepository;
+use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -22,6 +23,9 @@ class BaladesController extends AbstractController
     #[Route('/', name: 'app_balades_index', methods: ['GET'])]
     public function index(BaladesRepository $baladesRepository, ActivitesRepository $activitesRepository): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException('Vous n\'êtes pas autorisé à accéder à cette page');
+        }
         return $this->render('balades/index.html.twig', [
             'balades' => $baladesRepository->findAll(),
             'activites' => $activitesRepository->findAll(),
@@ -40,6 +44,9 @@ class BaladesController extends AbstractController
     #[Route('/new', name: 'app_balades_new', methods: ['GET', 'POST'])]
     public function new(Request $request, BaladesRepository $baladesRepository, SluggerInterface $slugger): Response
     {
+        if (!$this->isGranted('ROLE_USER')) {
+            throw $this->createAccessDeniedException('Vous n\'êtes pas autorisé à accéder à cette page');
+        }
        
         $balade = new Balades();
         $form = $this->createForm(BaladesType::class, $balade);
@@ -87,16 +94,35 @@ class BaladesController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_balades_show', methods: ['GET'])]
-    public function show(Balades $balade): Response
+ 
+    public function show(Balades $balade, Request $request, $id, ManagerRegistry $doctrine): Response
     {
+      
+       
+      $baladeId = $balade->getId();
+        // Traiter la soumission du formulaire
+        if ($request->isMethod('POST')) {
+            // Enregistrer le produit signalé dans une base de données ou un système de files d'attente
+            // ...
+            
+            // Rediriger l'utilisateur vers la page de détails du produit avec un message de confirmation
+            return $this->redirectToRoute('app_balade_detail', ['id' => $id, 'signaled' => true]);
+        }
+        
+       
         return $this->render('balades/show.html.twig', [
             'balade' => $balade,
+            'baladeId' => $baladeId,
+          
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_balades_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Balades $balade, BaladesRepository $baladesRepository): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException('Vous n\'êtes pas autorisé à accéder à cette page');
+        }
         $form = $this->createForm(BaladesType::class, $balade);
         $form->handleRequest($request);
 
@@ -115,10 +141,35 @@ class BaladesController extends AbstractController
     #[Route('/{id}', name: 'app_balades_delete', methods: ['POST'])]
     public function delete(Request $request, Balades $balade, BaladesRepository $baladesRepository): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException('Vous n\'êtes pas autorisé à accéder à cette page');
+        }
         if ($this->isCsrfTokenValid('delete'.$balade->getId(), $request->request->get('_token'))) {
             $baladesRepository->remove($balade, true);
         }
 
         return $this->redirectToRoute('app_balades_index', [], Response::HTTP_SEE_OTHER);
     }
+
+//     #[Route("/balade/{id}/signal", name: "app_signal_balade", methods: ['POST'])]
+    
+//    public function signalbalade(Request $request, $id)
+//    {
+//        // Récupérer le produit signalé à partir de la base de données
+//        $balade = $this->getDoctrine()->getRepository(balade::class)->find($id);
+       
+//        // Traiter la soumission du formulaire
+//        if ($request->isMethod('POST')) {
+//            // Enregistrer le produit signalé dans une base de données ou un système de files d'attente
+//            // ...
+           
+//            // Rediriger l'utilisateur vers la page de détails du produit avec un message de confirmation
+//            return $this->redirectToRoute('app_balade_detail', ['id' => $id, 'signaled' => true]);
+//        }
+       
+//        // Afficher le formulaire de signalement de produit
+//        return $this->render('balade/signal.html.twig', [
+//            'balade' => $balade,
+//        ]);
+//    }
 }
